requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.58.2
django-leaflet==0.29.0
docopt==0.6.2
GDAL==3.4.0
gpxpy==1.6.1
gunicorn==21.2.0
idna==3.4
Js2Py==0.74
oauthlib==3.2.2
packaging==23.2
pipwin==0.5.2
psycopg2==2.9.9
pycparser==2.21
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
whitenoise==6.6.0
